moved postUtils function to their own file. tests should pass now
